# scim-openapi.yaml
openapi: 3.0.3
info:
  title: SCIM 2.0 API
  version: 1.0.0
  description: >
    SCIM 2.0-compliant user provisioning API for Keycloak. All endpoints require OAuth2
    client credentials authentication (Bearer token). The client must be authenticated and
    the service account must have the realm role `scim-access`.
servers:
  - url: https://example.com/scim/v2
paths:
  /Users:
    get:
      summary: List users
      operationId: listUsers
      tags: [Users]
      description: >
        Returns a list of users in the current realm.

        Requires:
        - OAuth2 access token (Bearer token)
        - The client must be authenticated via client credentials
        - The service account must have the realm role `scim-access`
      security:
        - bearerAuth: []
      parameters:
        - name: filter
          in: query
          description: SCIM filter expression (e.g. `userName eq "alice@example.com"`)
          required: false
          schema:
            type: string
        - name: startIndex
          in: query
          schema:
            type: integer
        - name: count
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of users
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/UsersList'

    post:
      summary: Create user
      operationId: createUser
      tags: [Users]
      description: >
        Creates a new user in the realm.

        Requires:
        - OAuth2 access token (Bearer token)
        - The service account must have the realm role `scim-access`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
  /Users/{id}:
    get:
      summary: Get user by ID
      operationId: findUser
      tags: [ Users ]
      description: >
        Retrieves a user by SCIM ID.
        
        Requires:
        - OAuth2 access token (Bearer token)
        - The client must be authenticated via client credentials
        - The service account must have the realm role `scim-access`
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: SCIM user ID
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      summary: Update user
      operationId: updateUser
      tags: [ Users ]
      description: >
        Updates a user's full SCIM representation with the provided one.
        
        Requires:
        - OAuth2 access token (Bearer token)
        - The service account must have the realm role `scim-access`
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User replaced
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request (e.g., invalid schema)
        '404':
          description: User not found
        '409':
          description: Conflict (e.g., username already exists)
    delete:
      summary: Delete user
      operationId: deleteUser
      tags: [Users]
      description: >
        Deletes a user by ID.

        Requires:
        - OAuth2 access token (Bearer token)
        - The service account must have the realm role `scim-access`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted

    patch:
      summary: Update user
      operationId: patchUser
      tags: [Users]
      description: >
        Updates a user using SCIM PATCH semantics.

        Requires:
        - OAuth2 access token (Bearer token)
        - The service account must have the realm role `scim-access`
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/PatchRequest'
      responses:
        '200':
          description: User updated
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    UsersList:
      type: object
      properties:
        totalResults:
          type: integer
        startIndex:
          type: integer
        itemsPerPage:
          type: integer
        Resources:
          type: array
          items:
            $ref: '#/components/schemas/User'
    User:
      type: object
      required:
        - userName
        - schemas
      properties:
        id:
          type: string
        externalId:
          type: string
        userName:
          type: string
        displayName:
          type: string
        preferredLanguage:
          type: string
        name:
          type: object
          properties:
            givenName:
              type: string
            familyName:
              type: string
        emails:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
              primary:
                type: boolean
        active:
          type: boolean
        schemas:
          type: array
          items:
            type: string
        meta:
          $ref: '#/components/schemas/Meta'
...
    PatchRequest:
      type: object
      required:
        - schemas
        - Operations
      properties:
        schemas:
          type: array
          items:
            type: string
          example: [ "urn:ietf:params:scim:api:messages:2.0:PatchOp" ]
        Operations:
          type: array
          items:
            type: object
            required:
              - op
            properties:
              op:
                type: string
                enum: [ add, remove, replace ]
              path:
                type: string
                description: Attribute path to target
              value:
                description: >
                  Value to be applied. Can be a string, boolean, number, object, or array depending on the patch.
                type: object
...
    Meta:
      type: object
      required:
        - location
        - resourceType
      properties:
        location:
          type: string
          format: uri
          example: "https://example.com/v2/ServiceProviderConfig"
        resourceType:
          type: string
          example: "ServiceProviderConfig"
        created:
          type: string
          format: date-time
          example: "2010-01-23T04:56:22Z"
        lastModified:
          type: string
          format: date-time
          example: "2011-05-13T04:42:34Z"